+++ packages/sn-filter-pane/src/ext/index.ts
@@ -1,5 +1,7 @@
 import pp from './property-panel';
 import { IEnv } from '../types/types';
+import { createImportProperties, exportProperties } from './conversion';
+import { ExportFormat } from './conversion/types';
 
 export default function ext(env: IEnv) {
   return {
@@ -12,5 +14,7 @@ export default function ext(env: IEnv) {
       sharing: false,
       viewData: true,
     },
+    importProperties: (exportFormat:ExportFormat, initialProperties: EngineAPI.IGenericHyperCubeProperties) => createImportProperties(exportFormat, initialProperties),
+    exportProperties,
   };
 }