+++ CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog for sn-filter-pane and sn-listbox
 
+## [0.2.0](https://github.com/qlik-oss/sn-list-objects/compare/v0.1.0...v0.2.0) (2023-01-31)
+
+
+### Features
+
+* add chart conversion setup ([#42](https://github.com/qlik-oss/sn-list-objects/issues/42)) ([0da155e](https://github.com/qlik-oss/sn-list-objects/commit/0da155e039bbdd9dd45fabafcf521f1409c21b40))
+* add checkbox and histogram to property panel ([#36](https://github.com/qlik-oss/sn-list-objects/issues/36)) ([fd35401](https://github.com/qlik-oss/sn-list-objects/commit/fd35401cee4e528d0c71690508c6caa4eb24311b))
+* add dense mode ([#21](https://github.com/qlik-oss/sn-list-objects/issues/21)) ([2a1ee77](https://github.com/qlik-oss/sn-list-objects/commit/2a1ee7761b1874f85f78da115d5703303e8fc76a))
+* add dense option to each individual listbox ([#23](https://github.com/qlik-oss/sn-list-objects/issues/23)) ([743e423](https://github.com/qlik-oss/sn-list-objects/commit/743e42381f8e2565a56ad53d3e389ee0536a3e67))
+* add direction listboxOptions ([#33](https://github.com/qlik-oss/sn-list-objects/issues/33)) ([9e8ef94](https://github.com/qlik-oss/sn-list-objects/commit/9e8ef943b1f3d0affd05abec8084e0eb85286bb2))
+* add textAlign property ([#29](https://github.com/qlik-oss/sn-list-objects/issues/29)) ([9b5d3b3](https://github.com/qlik-oss/sn-list-objects/commit/9b5d3b308fd750a58d1bedc1a523d0e71964d856))
+* apply sense theme ([#34](https://github.com/qlik-oss/sn-list-objects/issues/34)) ([f578388](https://github.com/qlik-oss/sn-list-objects/commit/f57838821baa6f081a4db6d06e340872a88fb0a1))
+* import&export properties chart conversion ([#43](https://github.com/qlik-oss/sn-list-objects/issues/43)) ([2591967](https://github.com/qlik-oss/sn-list-objects/commit/259196757567554f7047672251720a6c2a21fa2f))
+* popover container ([#24](https://github.com/qlik-oss/sn-list-objects/issues/24)) ([e64a30c](https://github.com/qlik-oss/sn-list-objects/commit/e64a30c8cc1ead7a819345b9c88db0147dfef4f4))
+* use resource store ([#31](https://github.com/qlik-oss/sn-list-objects/issues/31)) ([c5b4b05](https://github.com/qlik-oss/sn-list-objects/commit/c5b4b0564f618ab457e4a6f51ffc9dcd727b1b91))
+
+
+### Bug Fixes
+
+* adapt to grid mode ([#28](https://github.com/qlik-oss/sn-list-objects/issues/28)) ([e1e156c](https://github.com/qlik-oss/sn-list-objects/commit/e1e156cba58a51e11110284466a8e4aa0512effb))
+* frequecyMode in sn-listbox ([#35](https://github.com/qlik-oss/sn-list-objects/issues/35)) ([26ec130](https://github.com/qlik-oss/sn-list-objects/commit/26ec130829886655ec04a929d76a7c1d4027b18e))
+* frequency enabled ([#30](https://github.com/qlik-oss/sn-list-objects/issues/30)) ([aa3bcf0](https://github.com/qlik-oss/sn-list-objects/commit/aa3bcf0d4596a021960ad0ac40ed03cb05a85473))
+* ts and lint errors/warnings ([#39](https://github.com/qlik-oss/sn-list-objects/issues/39)) ([e302159](https://github.com/qlik-oss/sn-list-objects/commit/e3021596c9a41f6f1be4b20a279a0d244f0c8256))
+* tweak grid layout ([#27](https://github.com/qlik-oss/sn-list-objects/issues/27)) ([5ba6090](https://github.com/qlik-oss/sn-list-objects/commit/5ba609023cb21232302c879eb0d7d913b106f36a))
+
 ## [0.1.0](https://github.com/qlik-oss/sn-list-objects/compare/v0.1.0-alpha.1...v0.1.0) (2022-11-14)
 
 ## [0.1.0-alpha.1](https://github.com/qlik-oss/sn-list-objects/compare/v0.1.0-alpha.0...v0.1.0-alpha.1) (2022-11-07)